moved graphchat component to ui folder
